services/consumer/requirements.txt
async-timeout==4.0.3
certifi==2024.7.4
charset-normalizer==3.3.2
dnspython==2.6.1
idna==3.8
pika==1.3.2
pymongo==4.8.0
python-dotenv==1.0.1
redis==5.0.8
requests==2.32.3
urllib3==2.2.2
